test: shift assignment creation
