+++ erpnext/hr/doctype/employee_checkin/test_employee_checkin.py
@@ -23,6 +23,8 @@ from erpnext.hr.doctype.employee_checkin.employee_checkin import (
 	mark_attendance_and_link_log,
 )
 from erpnext.hr.doctype.leave_application.test_leave_application import get_first_sunday
+from erpnext.hr.doctype.shift_assignment.test_shift_assignment import make_shift_assignment
+from erpnext.hr.doctype.shift_type.test_shift_type import setup_shift_type
 from erpnext.payroll.doctype.salary_slip.test_salary_slip import make_holiday_list
 
 
@@ -280,45 +282,3 @@ def make_checkin(employee, time=now_datetime()):
 		}
 	).insert()
 	return log
-
-
-def setup_shift_type(**args):
-	args = frappe._dict(args)
-	shift_type = frappe.new_doc("Shift Type")
-	shift_type.__newname = args.shift_type or "_Test Shift"
-	shift_type.start_time = args.start_time or "08:00:00"
-	shift_type.end_time = args.end_time or "12:00:00"
-	shift_type.holiday_list = args.holiday_list
-	shift_type.enable_auto_attendance = 1
-
-	shift_type.determine_check_in_and_check_out = (
-		args.determine_check_in_and_check_out
-		or "Alternating entries as IN and OUT during the same shift"
-	)
-	shift_type.working_hours_calculation_based_on = (
-		args.working_hours_calculation_based_on or "First Check-in and Last Check-out"
-	)
-	shift_type.begin_check_in_before_shift_start_time = (
-		args.begin_check_in_before_shift_start_time or 60
-	)
-	shift_type.allow_check_out_after_shift_end_time = args.allow_check_out_after_shift_end_time or 60
-
-	shift_type.save()
-
-	return shift_type
-
-
-def make_shift_assignment(shift_type, employee, start_date, end_date=None):
-	shift_assignment = frappe.get_doc(
-		{
-			"doctype": "Shift Assignment",
-			"shift_type": shift_type,
-			"company": "_Test Company",
-			"employee": employee,
-			"start_date": start_date,
-			"end_date": end_date,
-		}
-	).insert()
-	shift_assignment.submit()
-
-	return shift_assignment
